Some transaction fixes

+++ transaction/src/main/java/org/wildfly/httpclient/transaction/HttpRemoteTransactionHandle.java
@@ -80,10 +80,10 @@ class HttpRemoteTransactionHandle implements SimpleTransactionControl {
                 marshaller.start(output);
                 marshaller.writeInt(id.getFormatId());
                 final byte[] gtid = id.getGlobalTransactionId();
-                marshaller.writeByte(gtid.length);
+                marshaller.writeInt(gtid.length);
                 marshaller.write(gtid);
                 final byte[] bq = id.getBranchQualifier();
-                marshaller.writeByte(bq.length);
+                marshaller.writeInt(bq.length);
                 marshaller.write(bq);
                 marshaller.finish();
                 output.close();
@@ -143,10 +143,10 @@ class HttpRemoteTransactionHandle implements SimpleTransactionControl {
                 marshaller.start(output);
                 marshaller.writeInt(id.getFormatId());
                 final byte[] gtid = id.getGlobalTransactionId();
-                marshaller.writeByte(gtid.length);
+                marshaller.writeInt(gtid.length);
                 marshaller.write(gtid);
                 final byte[] bq = id.getBranchQualifier();
-                marshaller.writeByte(bq.length);
+                marshaller.writeInt(bq.length);
                 marshaller.write(bq);
                 marshaller.finish();
                 output.close();
